[ble] Fixes for py-pretty-check.
